Move vf-scripts task
